INS-451: rename event bus package
