+++ eventbus/response/common.go
