+++ network/servicenetwork/servicenetwork_test.go
@@ -25,7 +25,7 @@ import (
 
 	"github.com/insolar/insolar/configuration"
 	"github.com/insolar/insolar/core"
-	"github.com/insolar/insolar/messagerouter/message"
+	"github.com/insolar/insolar/eventbus/message"
 	"github.com/prometheus/common/log"
 	"github.com/stretchr/testify/assert"
 )
